FAC-573: Move NotificationCmd records to domain package
