+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/port/in/assessment/CreateAssessmentUseCase.java
@@ -7,7 +7,7 @@ import lombok.EqualsAndHashCode;
 import lombok.Value;
 import org.flickit.assessment.common.application.SelfValidating;
 import org.flickit.assessment.common.application.domain.notification.HasNotificationCmd;
-import org.flickit.assessment.common.application.domain.notification.NotificationCmd;
+import org.flickit.assessment.core.application.domain.notification.CreateAssessmentNotificationCmd;
 
 import java.util.UUID;
 
@@ -45,5 +45,6 @@ public interface CreateAssessmentUseCase {
         }
     }
 
-    record Result(UUID id, NotificationCmd notificationCmd) implements HasNotificationCmd {}
+    record Result(UUID id, CreateAssessmentNotificationCmd notificationCmd) implements HasNotificationCmd {
+    }
 }
